rename greetingFetcher and counterStore
