+++ starters/solidjs-tailwind/src/components/CounterExample/CounterDisplay.jsx
@@ -1,4 +1,4 @@
-import { count } from './cccounterStore';
+import { count } from './counterStore';
 
 const CounterDisplay = () => (
   <strong role="display-count" class="text-xl">
