+++ starters/solidjs-tailwind/src/components/FetchExample/Greeting.jsx
@@ -1,5 +1,5 @@
 import { createResource, Match, Switch } from 'solid-js';
-import greetingFetcher from './gggreetingFetcher';
+import greetingFetcher from './greetingFetcher';
 
 const Greeting = () => {
   const [message] = createResource(greetingFetcher());
